requirements.txt
flask 
twilio 
requests
Pillow 
pyzbar
gunicorn
